refactor(python): :recycle: Split code into seperate files
Code was in one long script. I've started the process to split these into seperate files.

This commit is completely broken.
