+++ speedflux/__init__.py
@@ -1,0 +1,2 @@
+from logs import log
+
